refactor(model): 重构多个模型类字段及服务接口
feat(Product): 添加@TableLogic注解支持逻辑删除
feat(OrderInvoice): 移除invoice_no字段，使用invoice_number
feat(ProductInventoryConfig): 添加updated_at字段
feat(UserInterestProfile): 添加profile_version字段
feat(CustomerServiceAgent): 添加last_activity_at字段
feat(ProductSearchKeyword): 添加点击和转化统计字段
refactor(UserReviewStats): 简化模型结构
refactor(ProductSku): 重构SKU模型字段
feat(InventoryTransaction): 完善库存交易模型注解
refactor(SellerProfileService): 重构卖家资料服务接口
feat(PaymentTransactionMapper): 添加单条支付统计查询
feat(ProductDetailMapper): 添加商品详情CRUD操作
feat(UserBehaviorLogMapper): 添加用户行为记录功能
refactor(PromotionActivityServiceImpl): 重构促销活动服务实现
refactor(UserReviewDetailServiceImpl): 重构评论详情服务实现
refactor(SellerProfileServiceImpl): 重构卖家资料服务实现

+++ daily-discover-product/src/main/java/com/dailydiscover/service/impl/UserReviewStatsServiceImpl.java
@@ -9,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.math.BigDecimal;
 
 @Service
 @Slf4j
@@ -24,7 +25,7 @@ public class UserReviewStatsServiceImpl extends ServiceImpl<UserReviewStatsMappe
     
     @Override
     public boolean updateUserReviewStats(Long userId, Integer totalReviews, Integer helpfulReviews, 
-                                        Integer unhelpfulReviews, BigDecimal averageRating) {
+                                        Integer unhelpfulReviews, Double averageRating) {
         UserReviewStats stats = getByUserId(userId);
         if (stats == null) {
             stats = new UserReviewStats();
